# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/activeUsers": {
      "get": {
        "description": "Returns all active users from the in-memory cache for the specified tenant within the given time range. Each user object includes identity, browser, device, platform, country, language, and role information.",
        "operationId": "ActiveUserController_getUsers",
        "parameters": [
          {
            "name": "start",
            "required": true,
            "in": "query",
            "description": "Start of the time range (ISO 8601 string or epoch milliseconds).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": true,
            "in": "query",
            "description": "End of the time range (ISO 8601 string or epoch milliseconds).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Tenant identifier used to scope the cache query.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "/activeUsers",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/sessionEvents": {
      "get": {
        "description": "Returns raw session-initialised events from the cache within the given time range for the specified tenant. Each entry contains `time` (ISO 8601), `eventId`, `identifyId`, `inferredLocation`, and `userType`.",
        "operationId": "SessionEventsController_getSessions",
        "parameters": [
          {
            "name": "start",
            "required": true,
            "in": "query",
            "description": "Start of the time range (ISO 8601 string or epoch milliseconds).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": true,
            "in": "query",
            "description": "End of the time range (ISO 8601 string or epoch milliseconds).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Tenant identifier used to scope the cache query.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "/sessionEvents",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/v2/activeUsers": {
      "get": {
        "description": "Supports optional server-side filtering via filtrex (https://github.com/cshaa/filtrex) and field projection. All string values in filter expressions must be double-quoted.",
        "operationId": "ActiveUserControllerV2_getUsersV2",
        "parameters": [
          {
            "name": "start",
            "required": true,
            "in": "query",
            "description": "Start of the time range (ISO 8601 string or epoch milliseconds).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": true,
            "in": "query",
            "description": "End of the time range (ISO 8601 string or epoch milliseconds).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Tenant identifier used to scope the cache query.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "filtrex filter expression applied to mapped user objects. Strings must be double-quoted.\n\n**Examples**\n- `country == \"Germany\"` — users from Germany\n- `platform == \"desktop\" and browser == \"Chrome\"` — desktop Chrome users\n- `device in (\"mobile\", \"tablet\")` — mobile and tablet users only\n- `roles != \"guest\"` — exclude guest users\n- `domain ~= \"siemens\"` — users whose email domain contains \"siemens\"",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "Comma-separated user fields to return. Valid values: id, roles, country, sessionId, language, platform, browser, device, domain. Omit to return all fields.\n\n**Examples**\n- `id,country,roles` — identity, location, and role breakdown\n- `id,platform,browser,device` — device and browser profile only\n- `id,domain,language` — domain and language breakdown",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "Field and optional direction to sort results by, in the format `field` or `field:asc` / `field:desc`. Direction defaults to `asc` when omitted. Valid field values match those listed in `fields`.\n\n**Examples**\n- `country:asc` — alphabetical by country\n- `domain:asc` — alphabetical by domain\n- `language:asc` — alphabetical by language\n- `platform:desc` — reverse-alphabetical by platform",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filtered and projected active user objects.",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "user-001",
                      "roles": [
                        "admin",
                        "developer"
                      ],
                      "country": "Germany",
                      "sessionId": "sess-abc123",
                      "language": "de",
                      "platform": "desktop",
                      "browser": "Chrome",
                      "device": "desktop",
                      "domain": "example.com"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "v2/activeUsers",
        "tags": [
          "v2"
        ]
      }
    },
...
    "/v2/sessionEvents": {
      "get": {
        "description": "Supports optional server-side filtering via filtrex (https://github.com/cshaa/filtrex) and field projection. All string values in filter expressions must be double-quoted.",
        "operationId": "SessionEventsControllerV2_getSessionsV2",
        "parameters": [
          {
            "name": "start",
            "required": true,
            "in": "query",
            "description": "Start of the time range (ISO 8601 string or epoch milliseconds).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": true,
            "in": "query",
            "description": "End of the time range (ISO 8601 string or epoch milliseconds).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Tenant identifier used to scope the cache query.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "filtrex filter expression applied to mapped session event objects. Strings must be double-quoted.\n\n**Examples**\n- `userType == \"USER\"` — regular authenticated users only\n- `userType != \"VISITOR\"` — exclude anonymous visitors\n- `userType in (\"USER\", \"LEAD\")` — users and leads\n- `accountId == \"acc-001\"` — sessions for a specific account\n- `eventType == \"SESSION_STARTED\"` — session start events only",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "Comma-separated session event fields to return. Valid values: id, identifyId, sessionId, accountId, date, propertyKey, eventType, remoteHost, location, userType, globalContext. Omit to return all fields.\n\n**Examples**\n- `id,identifyId,accountId,date` — minimal identity and timestamp\n- `id,userType,location,date` — geographic and user type analysis\n- `id,identifyId,sessionId,userType` — session tracking with user type",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "Field and optional direction to sort results by, in the format `field` or `field:asc` / `field:desc`. Direction defaults to `asc` when omitted. Valid field values match those listed in `fields`.\n\n**Examples**\n- `date:desc` — most recent sessions first\n- `date:asc` — oldest sessions first\n- `accountId:asc` — alphabetical by account\n- `userType:asc` — alphabetical by user type",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filtered and projected session event objects.",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "se-001",
                      "identifyId": "user-001",
                      "sessionId": "sess-abc123",
                      "accountId": "acc-001",
                      "date": "2024-01-15T10:30:00.000Z",
                      "propertyKey": "APX-12345",
                      "eventType": "SESSION_STARTED",
                      "remoteHost": "203.0.113.10",
                      "userType": "USER",
                      "location": {
                        "countryName": "Germany",
                        "regionName": "Bavaria"
                      },
                      "globalContext": []
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "v2/sessionEvents",
        "tags": [
          "v2"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gainsight Sync Microservice API",
    "description": "A NestJS microservice that periodically pulls Gainsight PX telemetry (custom events, session events, page views, active-user metrics) from configured tenant domains, caches the data in memory per tenant using a date-sorted array with binary-search range queries, and exposes REST endpoints consumed by dashboards such as Grafana.\n\nAll data endpoints require `Authorization: Basic <base64(user:password)>` and a `tenantId` query parameter. Public endpoints: `/health`, `/lastRun`, `/llms.txt`, `/llms-full.txt`.\n\nInteractive docs: https://gainsight.eu-latest.cumulocity.com/apps/gainsight-c8y-openapi/index.html",
    "version": "0.2.2",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://gainsight.eu-latest.cumulocity.com/service/gainsight-sync-ms"
    }
  ],
  "components": {
    "securitySchemes": {
      "basic": {
        "type": "http",
        "scheme": "basic"
      }
    },
...
  }
}